Fixed a python setup issue + changed essential Homebrew formulae and casks

+++ system-setup/macos-tasks/g-setup-python.nu
@@ -1,13 +1,13 @@
 export const order = 7
 
 def setup_python [] {
-  use ../utils/utils.nu ensure_homebrew_package
+  use ../utils/utils.nu [ensure_homebrew_package, command_exists]
   print "Setting up Python..."
 
   mkdir --verbose $"($nu.home-path)/.local/share/pyenv"
   with-env { PYENV_ROOT: "~/.local/share/pyenv", PYTHON_CONFIGURE_OPTS: "--enable-optimizations --with-lto", PYTHON_CFLAGS: "-march=native -mtune=native" } {
     ensure_homebrew_package "pyenv"
-    if not (which python | get path.0 | str contains "pyenv") {
+    if not ((command_exists "python") and (which python | get path.0 | str contains "pyenv")) {
       print "Installing the latest version of Python 3 with pyenv..."
       pyenv install --skip-existing 3
       pyenv global 3
